src/requirements.txt
auto_mix_prep==0.2.0
catboost==1.0.4
category_encoders==2.3.0
dataclasses_json==0.5.6
inflection==0.5.1
mlplatform_lib==8.4.1.2.13.0.0
# numpy==1.22.4
numpy==1.20.3
#pandas==1.5.2
pandas==1.3.5
tables==3.7.0
PyYAML==6.0
scikit_learn==1.0.2
# scikit_learn==1.1.1
torch==1.13.0